Apply role colors to displayed usernames
