+++ src/api/manApi.ts
@@ -453,6 +453,7 @@ export type ForumThread = {
   id: number;
   title: string;
   author_username?: string | null;
+  author_role?: UserRole | null;
   author_avatar_url?: string | null;
   author_avatar_preset?: AvatarPreset | null;
   created_at: string;
@@ -466,6 +467,7 @@ export type ForumThread = {
 export type ForumPost = {
   id: number;
   author_username?: string | null;
+  author_role?: UserRole | null;
   author_avatar_url?: string | null;
   author_avatar_preset?: AvatarPreset | null;
   content_markdown: string;
